requirements_czsl.txt
absl-py==1.3.0
-e git+https://github.com/altair-viz/altair.git@8a71da274ca9bddf4ff29d7862b890e7927306da#egg=altair
apt-clone==0.2.1
apturl==0.5.2
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.2.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cmake==3.27.7
colorama==0.4.6
comm==0.1.2
command-not-found==0.3
commonmark==0.9.1
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.4
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
easydict==1.10
executing==1.2.0
fasteners==0.14.1
filelock==3.13.1
future==0.18.2
google-auth==2.15.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.51.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.19.4
ipython==8.7.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==17.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
meld==3.20.4
mock==4.0.3
model-index==0.1.11
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.2.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
ordered-set==4.1.0
packaging==22.0
PAM==0.4.2
pandas==2.0.3
paramiko==2.9.3
parso==0.8.3
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.6.0
prompt-toolkit==3.0.36
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycups==2.0.1
Pygments==2.13.0
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.19.2
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
QtPy==2.0.0
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
ssh-import-id==5.11
stack-data==0.6.2
sympy==1.12
systemd-python==234
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
toolz==0.12.0
torch==2.0.1
torch-tb-profiler==0.4.0
torchvision==0.15.2
tornado==6.2
traitlets==5.8.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
Werkzeug==2.2.2
xdg==5
xkit==0.0.0
zipp==1.0.0